Created a new questions page to display questions. Added click feedback for bookmark and ensured font consistency
