+++ src/styles/form.css
@@ -59,7 +59,7 @@ textarea:focus {
 @media (min-width:1440px) {
 
     textarea {
-        height: 60vh;
+        height: 56vh;
     }
 
 }
